Fix Bookcreator, Delete and Rename page plugins
Bookcreator relies on "plugin_bookcreator__addtobook" class into "a" link to work. It's added by default by "getListItems(); ?>" function, but as Argon Alt builds a custom menu, it's not appending this class automatically to the "a" link and the bookcreator link doesn't work. It also fix compatibility issues with other plugins that relies on classes to properly function, as "Delete page" and "Rename page" plugins.